requirements.txt
google-generativeai
python-dotenv
PyPDF2
numpy
pandas
scikit-learn
tqdm
streamlit
spacy
rapidfuzz
